requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
multidict==6.7.0
openai==2.3.0
propcache==0.4.1
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.41
stripe==13.0.1
tqdm==4.67.1
twilio==9.8.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
gunicorn==21.2.0

google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
# Security Enhancement Packages
psycopg2-binary==2.9.9  # PostgreSQL adapter
cryptography==42.0.5  # Encryption utilities
bcrypt==4.1.2  # Password hashing
flask-talisman==1.1.0  # HTTPS enforcement and security headers
flask-limiter==3.5.0  # Rate limiting
flask-wtf==1.2.1  # Form validation and CSRF protection
wtforms==3.1.2  # Input validation
python-dateutil==2.8.2  # Date utilities for retention policies
boto3==1.34.51  # AWS S3 for backups

